complete todo lo requerido para hacer un register. me encontre con un error de db.hibernate_sequense table not found, la unica manera de solucinarlo fue agregar una nueva notacion en el application properties lo que sugiere que el error se debe a un upgrade en las version de hibernate durante el proyecto

+++ src/main/java/ar/com/ada/api/aladas/services/VueloService.java
@@ -100,7 +100,7 @@ public class VueloService {
          * else return false;
          */
 
-        return vuelo.getAeropuertoDestino() != vuelo.getAeropuertoOrigen();
+        return vuelo.getAeropuertoDestino().intValue() != vuelo.getAeropuertoOrigen().intValue(); // mmodificar a equals o agregar el .intValue()
 
     }
